Starting to rework the entire codebase to be context friendly. Annotations is next after one successful install.
