+++ .gitignore
@@ -1,2 +1,4 @@
 **/**__pycache__
 SAFETY_LOCK
+**/**old.*
+**/**.img
