+++ exceptions.py
@@ -1,0 +1,4 @@
+class RequirementError(BaseException):
+	pass
+class DiskError(BaseException):
+	pass
